Changed custom creep models to npc_units
Tower Prefabs were hardcoded and read-only so new ones were made to be
able to edit them.

The Treant tower needed cosmetics so instead of using a npc_dota_tower
in the prefab, a npc_dota_base was used. dota_tower_creature is an
ability to take care of the tower properties like magic immunity and
throne invuln count

Bugs: Backdoor protection is not being disabled, Dire Throne is still
invulnerable when both towers are killed (feel free to add the
dota_tower_creature ability, it should work for towers too, as long as
its skilled up)